link directly to the issue list
I was confused when I was trying to find the current proposal doc for tomorrow's meeting, because this link sent me to the repository itself (and assumed I would have read the whole sentence and thus know to follow the link to the issues tab there).

But we might as well just link directly to the issues tab. that's what I'm doing in this change.