Link to Github view of expert map
Right now, clicking the link to the expert map downloads it to your machine. This isn't very convenient for viewing (or editing), so this makes it link to the file on Github.

Unfortunately this means we'll have to update the path anytime we move things around. I'm not sure how to get around that.