Here is my attempt to address #334.
The steering-meeting changes are, I think, obviously better.

The only change I"m not sure about is the change so that "Compiler team
procedures" links to the specific compiler page on forge. The main problem I
have with that is that page itself does not actually document any procedures,
nor even *link* to subpages with such documentation.

 * However, I justify this change by saying that I think in an ideal world, each
   main page would have links to each of its subpages, because often the titles
   alone in the Table-of-Contents on the left-hand side are not really enough
   for a reader who arrives on the forge without any experience or context for
   what they are looking for.